Реализация метода получения сообщения по WS соединению.

+++ internal/models/chat.go
@@ -1,6 +1,6 @@
 package models
 
 type Chat struct {
-	ID   uint   `json:"chat_id"`
+	ID   uint64 `json:"chat_id"`
 	Name string `json:"name"`
 }
